refactor(schematics): add schematic for ts import migration from simpl to siemens
-added comprehensive import mappings for Simpl to Siemens component migration
