+++ projects/element-ng/schematics/simpl-siemens-migration/mappings/index.ts
@@ -2,9 +2,4 @@
  * Copyright (c) Siemens 2016 - 2025
  * SPDX-License-Identifier: MIT
  */
-export * from './charts-ng-mappings.js';
-export * from './dashboards-ng-mappings.js';
-export * from './element-ng-mappings.js';
-export * from './element-translate-ng-mappings.js';
-export * from './maps-ng-mappings.js';
 export * from './style-mappings.js';
