+++ projects/element-ng/schematics/ts-import-to-siemens-migration/index.ts
@@ -5,6 +5,7 @@
 import { Tree, SchematicContext, Rule, chain } from '@angular-devkit/schematics';
 import * as ts from 'typescript';
 
+import { Migrations } from '../simpl-siemens-migration/model.js';
 import { getImportNodes, getSymbols, discoverSourceFiles } from '../utils/index.js';
 import {
   CHARTS_NG_MAPPINGS,
@@ -13,7 +14,13 @@ import {
   ELEMENT_TRANSLATE_NG_MAPPINGS,
   MAPS_NG_MAPPINGS
 } from './mappings/index.js';
-import { Migrations } from './model.js';
+
+export const tsImportMigration = (_options: { path: string }): Rule => {
+  return (tree: Tree, context: SchematicContext) => {
+    context.logger.info('🚀 Starting Simpl to Siemens migration...');
+    return tsImportMigrationRule(_options)(tree, context);
+  };
+};
 
 /**
  * Creates a migration rule that updates import statements in TypeScript files.
@@ -27,10 +34,10 @@ import { Migrations } from './model.js';
  *
  * @example
  * ```typescript
- * const migrationRule = importMigrationRule({ path: 'some-path' });
+ * const migrationRule = tsImportMigrationRule({ path: 'some-path' });
  * ```
  */
-export const importMigrationRule = (_options: { path: string }): Rule => {
+export const tsImportMigrationRule = (_options: { path: string }): Rule => {
   return (tree: Tree, context: SchematicContext) => {
     const rules: Rule[] = [];
     context.logger.info('📦 Migrating TypeScript imports...');
