+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/charts-ng-mappings.ts
