+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/dashboards-ng-mappings.ts
