+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/element-ng-mappings.ts
