+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/element-translate-ng-mappings.ts
