+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Copyright (c) Siemens 2016 - 2025
+ * SPDX-License-Identifier: MIT
+ */
+export * from './charts-ng-mappings.js';
+export * from './dashboards-ng-mappings.js';
+export * from './element-ng-mappings.js';
+export * from './element-translate-ng-mappings.js';
+export * from './maps-ng-mappings.js';
