+++ projects/element-ng/schematics/ts-import-to-siemens-migration/mappings/maps-ng-mappings.ts
